+++ lib/rubocop/cop/rspec/conditional_in_let.rb
@@ -25,8 +25,8 @@ module RuboCop
       #   end
       #
       class ConditionalInLet < ::RuboCop::Cop::Base
-        MSG = 'Do not put conditional logic in a `let` — use separate contexts.'
         LET_METHODS = %i[let let!].freeze
+        MSG = 'Do not put conditional logic in a `let` — use separate contexts.'
 
         def on_block(node)
           return unless let_block?(node)