+++ lib/rubocop/cop/rspec/factory_bot_in_before.rb
@@ -17,8 +17,8 @@ module RuboCop
       #   let(:user) { create(:user) }
       #
       class FactoryBotInBefore < ::RuboCop::Cop::Base
-        MSG = 'Do not create objects in `before` — use `let` for object creation.'
         CREATION_METHODS = %i[create build create_list build_list].freeze
+        MSG = 'Do not create objects in `before` — use `let` for object creation.'
 
         def on_block(node)
           return unless before_block?(node)